Algebraic data types
This adds simple algebraic data types. Each type consists of a set of
constructors, which each have an associated arity.

ADTs can be pattern matched, and pattern matching against constructors
belonging to a different type is a runtime type error.

Previously vectors served the same purpose, and vector pattern matching
didn't require vectors to have the same shape for this reason. With
this change, pattern matching vectors against a pattern with a different
shape is now a type error.