New parser
This replaces nom with a handrolled parser. The main purpose of this
change is to associate ranges in the source code with tokens,
expressions, etc. for better error reporting.

Currently these ranges are not persisted in Expr itself, and are only
used for displaying readable errors while parsing, and not at runtime
(future work).