Add array type distinct from tuples
This adds a new type for arrays. Internally, this is represented the
same as tuples, but the operations supported are different:

1) Array supports getting the element at an index, getting the length of
   the array, and concatenating two arrays.
2) Tuples can be used as patterns.

The idea of this is basically that tuples are like anonymous data
constructors, but represent a specific type and we shouldn't be able to
do things like query their length or concatenate them.

Arrays on the other hand represent a homogeneous list of values which
should probably be of the same type.