Add pretty-printer
This adds a new command "format" which pretty-prints a source file.

The pretty-printer converts the file to an Sexpr, and then prints it -
no formatting from the original, such as whitespace, is preserved.
However, quoted tokens are output unmodified (for example, "\n" won't be
replaced by an actual newline or vice versa).