Add support for loops
A loop expression defines a tail recursive function which is immediately
applied. This can be executed iteratively rather than recursively, and
avoid issues of stack overflows.

The advantage of having this as part of the language syntax rather than
just guaranteeing that tail recursive functions get optimized is that
it's clearer what will have TCO and what won't, and in particular, the
compiler can enforce that loops are actually defined in a tail recursive
way. It will also allow implementing stack traces for function calls,
without worrying about TCO removing entries from the stack if they
happen to be tail recursive.

Examples are updated to use loops where appropriate.

Currently there is no static checking that next (the keyword to continue
in a loop) only occurs in valid positions. This will be added in a
subsequent commit.