New syntax
This changes the syntax to a whitespace-significant one. The existing
syntax is used within a line, but parentheses no longer extend between
lines, and instead grouping over multiple lines is indicated by
indentation.

For example, what would be written like this:

; let $sum (
  fn $xs
; match xs
    (cons $x $xs) (add x; sum xs)
    nil 0
)
; sum some_list

Could now be written like this:

let $sum
  fn $xs
  match xs
    case; cons $x $xs
    add x (sum xs)
  , case nil 0
sum some_list

Subsequent lines at the same indentation are treated as if there are ";"
between them, so the later lines are children of the earlier ones. A
comma can be used if a line is a sibling of the previous ones, rather
than a child.

With the new syntax, the first element of each list has to be an atom
(i.e., it can't be a list itself). These are no longer really
S-expressions, and the type has been renamed to "Phrase", with the first
element being called "head". This plays better with the whitespace-based
syntax and mostly matches how the language works already - some
differences are adding "case" and "var" keywords for match cases and
loop variable definitions respectively.

A function which is not an atom can't be applied with the new syntax.
This is not really a problem, since functions are usually bound to
variables, and if necessary an "apply" function can be used.