Partial application
This implements currying/partial application of functions. This wasn't
implemented before to allow functions of zero arguments (whereas with
currying, all functions have one argument). However with the new syntax,
it's ambiguous whether a function of zero arguments is being invoked, or
just referred to, so this is not really something that can be supported
anyway.

In cases where a function of zero arguments makes sense, such as with
side effects or to defer computation, a function with a dummy argument
can be used instead.